+++ src/components/chats/ChatboxControls.scss
@@ -1,4 +1,4 @@
-@import "/src/layouts/colors.scss";
+@import "/src/components/layouts/colors.scss";
 
 .chatbox__header__avatar {
   flex: 0 0 3rem;
