+++ src/components/chats/ChatboxMessage.scss
@@ -1,4 +1,4 @@
-@import "/src/layouts/colors.scss";
+@import "/src/components/layouts/colors.scss";
 
 .chatbox__date {
   margin: 2rem 0;
