+++ src/components/chats/ChatlistItem.scss
@@ -1,4 +1,4 @@
-@import "/src/layouts/colors.scss";
+@import "/src/components/layouts/colors.scss";
 
 .chatlist__item {
   cursor: pointer;
