+++ src/components/forms/StandardButton.scss
@@ -1,4 +1,4 @@
-@import "/src/layouts/colors.scss";
+@import "/src/components/layouts/colors.scss";
 
 .form__button_standard {
   display: inline-block;
