+++ src/components/forms/StandardInput.scss
@@ -1,4 +1,4 @@
-@import "/src/layouts/colors.scss";
+@import "/src/components/layouts/colors.scss";
 
 .form__input__field_standard {
   padding: 1rem 0 0.7rem;
