+++ src/components/forms/common.scss
@@ -1,4 +1,4 @@
-@import "/src/layouts/colors.scss";
+@import "/src/components/layouts/colors.scss";
 
 .form {
   width: inherit;
