+++ src/components/layouts/Narrow.scss
