+++ src/components/layouts/Wide.scss
