--- src/modules/ProtoPages.js
@@ -1,14 +1,0 @@
-/*
-# ProtoPagesJS v0.0.4a
-
-## Dynamic template builder for the web powered by JavaScript
-
-https://github.com/sapomaro/ProtoPagesJS
-
-*/
-
-import {EventBus} from '/src/modules/ProtoPages-EventBus.js';
-import {ProtoView} from '/src/modules/ProtoPages-View.js';
-import {ProtoBlock} from '/src/modules/ProtoPages-Component.js';
-
-export {EventBus, ProtoView, ProtoBlock};
