+++ src/pages/500.html
@@ -1,17 +1,14 @@
 <!DOCTYPE html>
 <script type="module">
-	import {ProtoView} from '/src/modules/ProtoPages.js';
-
-  import {WideLayout} from '/src/layouts/Wide.js';
+  import {WideLayout} from '/src/components/layouts/Wide.js';
   import {ErrorStatusMsg} from '/src/components/ErrorStatusMsg.js';
 
-  const view = new ProtoView(WideLayout);
-  const context = view.use({});
-  
-  context.title = '500';
-  context.comment = 'Уже фиксим';
-  context.contents = ErrorStatusMsg; 
+  const view = new WideLayout({
+    title: '500',
+    comment: 'Уже фиксим',
+    contents: ErrorStatusMsg,
+  });
 
-  view.render();
+  view.renderToBody(); 
 
 </script>
