+++ src/pages/auth.html
@@ -1,20 +1,19 @@
 <!DOCTYPE html>
 <script type="module">
-	import {ProtoView} from '/src/modules/ProtoPages.js';
-  import {NarrowLayout} from '/src/layouts/Narrow.js';
+  import {NarrowLayout} from '/src/components/layouts/Narrow.js';
   import {Form} from '/src/components/forms/Form.js';
   import {StandardInput as Input} from '/src/components/forms/StandardInput.js';
   import {StandardButton as Button} from '/src/components/forms/StandardButton.js';
   import {StandardLink as Link} from '/src/components/forms/StandardLink.js';
   
-  const view = new ProtoView(NarrowLayout);
-  const context = view.use({Form, Input, Button, Link});
+  const view = new NarrowLayout({
+    title: 'Вход',
+    Form, Input, Button, Link
+  });
+  
+  view.props.contents = `%{ Form({ "name": "auth", "action": "chats.html" }) }%`;
 
-  context.title = 'Вход';
-
-  context.contents = `%{ Form({ "name": "auth", "action": "chats.html" }) }%`;
-
-  context.fieldset = () => `
+  view.props.fieldset = () => `
     <h1 class="container__header">%{title}%</h1>
     %{ Input({ "name": "login", "type": "text", "label": "Логин" }) }%
     %{ Input({ "name": "password", "type": "password", "label": "Пароль" }) }%
@@ -23,6 +22,6 @@
     %{ Link({ "url": "register.html", "label": "Нет аккаунта?" }) }%
   `;
 
-  view.render();
+  view.renderToBody();
 
 </script>
