+++ src/pages/chats.html
@@ -1,7 +1,6 @@
 <!DOCTYPE html>
 <script type="module">
-	import {ProtoView} from '/src/modules/ProtoPages.js';
-	import {ChatsLayout} from '/src/layouts/Chats.js';
+	import {ChatsLayout} from '/src/components/layouts/Chats.js';
   import {Form} from '/src/components/forms/Form.js';
   import {StandardInput as Input} from '/src/components/forms/StandardInput.js';
   import {StandardButton as Button} from '/src/components/forms/StandardButton.js';
@@ -11,44 +10,42 @@
   import {ChatboxHeader} from '/src/components/chats/ChatboxHeader.js';
   import {ChatboxFooter} from '/src/components/chats/ChatboxFooter.js';
   import {ChatboxTextarea} from '/src/components/chats/ChatboxTextarea.js';
+  import {JSONWrapper} from '/src/modules/Utils.js';
 
-  const view = new ProtoView(ChatsLayout);
-  const context = view.use({
+  const view = new ChatsLayout({
+    title: 'Чаты',
+    user: 'Собеседник',
     Form, Input, Button,
     ChatlistItem, ChatlistControls,
     ChatboxMessage, ChatboxHeader, ChatboxFooter, ChatboxTextarea
   });
 
-  context.title = 'Чаты';
-
-  context.user = 'Собеседник';
-
-  const chats = ProtoView.JSON.stringify([
+  const chats = JSONWrapper.stringify([
     { user: 'Андрей', quote: 'Изображение', when: '10:49', unreads: 2, active: true },
     { user: 'Илья', quote: 'Друзья, у меня для вас особенный выпуск новостей!...', when: '15:12', unreads: 4222 },
     { user: 'Design Destroyer', quote: 'В 2008 году художник Jon Rafman  начал собирать...', when: 'Пн', unreads: 0 },
   ]);
-  const messages = ProtoView.JSON.stringify([
+  const messages = JSONWrapper.stringify([
     { type: 'inc', when: '11:56', text: `Привет! Смотри, тут всплыл интересный кусок лунной космической истории — НАСА в какой-то момент попросила Хассельблад адаптировать модель SWC для полетов на Луну. Сейчас мы все знаем что астронавты летали с моделью 500 EL — и к слову говоря, все тушки этих камер все еще находятся на поверхности Луны, так как астронавты с собой забрали только кассеты с пленкой.
 
 Хассельблад в итоге адаптировал SWC для космоса, но что-то пошло не так и на ракету они так никогда и не попали. Всего их было произведено 25 штук, одну из них недавно продали на аукционе за 45000 евро.` },
     { type: 'out', when: '12:00', status: 1, text: 'Круто!' }
   ]);
 
-  context.Chatlist = () => `
+  view.props.Chatlist = () => `
     %{ ChatlistControls }%
     <ul>
       %{ ChatlistItem(${chats}...) }%
     </ul>
   `; 
 
-  context.ChatboxBody = () => `
+  view.props.ChatboxBody = () => `
     <div class="chatbox__date">19 июня (вс)</div>
 
     %{ ChatboxMessage(${messages}...) }%
   `;
 
-  context.popup = new Form({
+  view.props.popup = new Form({
     name: "addUser",
     fieldset: () => `
       <h1 class="container__header">Добавить пользователя</h1>
@@ -59,6 +56,6 @@
     `,
   });
 
-  view.render();
+  view.renderToBody();
 
 </script>
