+++ src/pages/profile_edit.html
@@ -1,19 +1,20 @@
 <!DOCTYPE html>
 <script type="module">
-  import {ProtoView} from '/src/modules/ProtoPages.js';
-  import {WideSideLayout} from '/src/layouts/Wide+Side.js';
+  import {WideSideLayout} from '/src/components/layouts/Wide+Side.js';
   import {Form} from '/src/components/forms/Form.js';
   import {StandardButton as Button} from '/src/components/forms/StandardButton.js';
   import {RowInput as Input} from '/src/components/forms/RowInput.js';
   import {RowLink as Link} from '/src/components/forms/RowLink.js';
   import {AvatarPic} from '/src/components/forms/Avatar.js';
+  import {JSONWrapper} from '/src/modules/Utils.js';
 
-  const view = new ProtoView(WideSideLayout);
-  const context = view.use({Form, Button, Input, Link, AvatarPic});
+  const view = new WideSideLayout({
+    title: 'Изменить данные',
+    backButtonLink: 'profile.html',
+    Form, Button, Input, Link, AvatarPic
+  });
 
-  context.title = 'Изменить данные';
-  context.backButtonLink = 'profile.html';
-  context.contents = `%{ Form({ "name": "profile", "action": "" }) }%`;
+  view.props.contents = `%{ Form({ "name": "profile", "action": "" }) }%`;
   
   const userData = {
     email: 'pochta@yandex.ru',
@@ -37,9 +38,9 @@
     input.value = userData[input.name];
   }
 
-  inputs = ProtoView.JSON.stringify(inputs);
+  inputs = JSONWrapper.stringify(inputs);
 
-  context.fieldset = () => `
+  view.props.fieldset = () => `
     %{ AvatarPic }%
     %{ Input(${inputs}...) }%
     <br><br><br>
@@ -47,6 +48,6 @@
     <br><br>
   `;
 
-  view.render();
+  view.renderToBody();
 
 </script>
