+++ src/pages/profile_newpass.html
@@ -1,20 +1,20 @@
 <!DOCTYPE html>
 <script type="module">
-  import {ProtoView} from '/src/modules/ProtoPages.js';
-  import {WideSideLayout} from '/src/layouts/Wide+Side.js';
+  import {WideSideLayout} from '/src/components/layouts/Wide+Side.js';
   import {Form} from '/src/components/forms/Form.js';
   import {StandardButton as Button} from '/src/components/forms/StandardButton.js';
   import {RowInput as Input} from '/src/components/forms/RowInput.js';
   import {AvatarPic} from '/src/components/forms/Avatar.js';
 
-  const view = new ProtoView(WideSideLayout);
-  const context = view.use({Form, Button, Input, AvatarPic});
+  const view = new WideSideLayout({
+    title: 'Изменить пароль',
+    backButtonLink: 'profile.html',
+    Form, Button, Input, AvatarPic
+  });
 
-  context.title = 'Изменить пароль';
-  context.backButtonLink = 'profile.html';
-  context.contents = `%{ Form({ "name": "changepassword", "action": "profile.html" }) }%`;
+  view.props.contents = `%{ Form({ "name": "changepassword", "action": "profile.html" }) }%`;
 
-  context.fieldset = () => `
+  view.props.fieldset = () => `
     %{ AvatarPic }%
     %{ Input({
       "label": "Старый пароль", "name": "oldPassword", "type": "password",
@@ -33,6 +33,6 @@
     <br><br>
   `;
 
-  view.render();
+  view.renderToBody();
 
 </script>
