+++ src/pages/register.html
@@ -1,18 +1,18 @@
 <!DOCTYPE html>
 <script type="module">
-	import {ProtoView} from '/src/modules/ProtoPages.js';
-  import {NarrowLayout} from '/src/layouts/Narrow.js';
+  import {NarrowLayout} from '/src/components/layouts/Narrow.js';
   import {Form} from '/src/components/forms/Form.js';
   import {StandardInput as Input} from '/src/components/forms/StandardInput.js';
   import {StandardButton as Button} from '/src/components/forms/StandardButton.js';
   import {StandardLink as Link} from '/src/components/forms/StandardLink.js';
+  import {JSONWrapper} from '/src/modules/Utils.js';
 
-  const view = new ProtoView(NarrowLayout);
-  const context = view.use({Form, Input, Button, Link});
+  const view = new NarrowLayout({
+    title: 'Регистрация',
+    Form, Input, Button, Link
+  });
 
-  context.title = 'Регистрация';
-
-  const inputs = ProtoView.JSON.stringify([
+  const inputs = JSONWrapper.stringify([
     { name: 'email', type: 'email', label: 'Почта' },
     { name: 'login', type: 'text', label: 'Логин' },
     { name: 'first_name', type: 'text', label: 'Имя' },
@@ -21,10 +21,10 @@
     { name: 'password', type: 'password', label: 'Пароль' },
     { name: 'password2', type: 'password', label: 'Пароль (ещё раз)' }
   ]);
-  
-  context.contents = `%{ Form({ "name": "reg", "action": "chats.html" }) }%`;
 
-  context.fieldset = () => `
+  view.props.contents = `%{ Form({ "name": "reg", "action": "chats.html" }) }%`;
+
+  view.props.fieldset = () => `
     <h1 class="container__header">%{title}%</h1>
     %{ Input(${inputs}...) }%
     <br><br>
@@ -32,6 +32,6 @@
     %{ Link({ "url": "auth.html", "label": "Войти" }) }%
   `;
 
-  view.render();
+  view.renderToBody();
     
 </script>
