+++ src/pages/test.html
@@ -1,15 +1,13 @@
 <!DOCTYPE html>
 <script type="module">
-	import {ProtoView} from '/src/modules/ProtoPages.js';
-  import {WideLayout} from '/src/layouts/Wide.js';
+  import {WideLayout} from '/src/components/layouts/Wide.js';
 	import {AjaxRequest} from '/src/modules/AjaxService.js';
 
-  const view = new ProtoView(WideLayout);
-  const context = view.use({});
+  const view = new WideLayout({
+    title: 'Тест XHR'
+  });
 
-  context.title = 'Тест XHR';
-
-  context.contents = () => `
+  view.props.contents = () => `
     <h1>%{title}%:</h1>
     <textarea>%{result1}%</textarea>
     <textarea>%{result2}%</textarea>
@@ -20,16 +18,16 @@
   })
   .then(({response}) => {
     console.log(response);
-    context.result1 = 'success';
+    view.props.result1 = 'success';
   })
   .catch(({error}) => {
     console.warn(error);
-    context.result1 = 'error';
+    view.props.result1 = 'error';
   })
   .finally(() => {
     console.log('finally');
-    context.result1 += ' finally';
-    view.render();
+    view.props.result1 += ' finally';
+    view.refresh();
   });
   
   const ajax2 = AjaxRequest({
@@ -37,18 +35,18 @@
   })
   .then(({response}) => {
     console.log(response);
-    context.result2 = 'success';
+    view.props.result2 = 'success';
   })
   .catch(({error}) => {
     console.warn(error);
-    context.result2 = 'error';
+    view.props.result2 = 'error';
   })
   .finally(() => {
     console.log('finally');
-    context.result2 += ' finally';
-    view.render();
+    view.props.result2 += ' finally';
+    view.refresh();
   });
   
-  view.render();
+  view.renderToBody(); 
 
 </script>
